BREAKING: Task groups are now exposed as part of the API
This makes anyio-sqlite more in line with idiomatic AnyIO/Trio while still
preserving decent performance.

- BREAKING: anyio_sqlite.connect() can only be used as a context manager;
it is a convenience wrapper around creating a task group and passing the task
group to the connection
- BREAKING: anyio_sqlite.Connection requires a task group.
- BREAKING: The default `iter_chunk_size` is now 128. 64 was fairly slow even on
asyncio due to AnyIO overhead.
- feat: added class method `anyio_sqlite.Connection.connect` to create a connection
without a context manager. A task group is required.
- feat: exiting the context manager without error commits the transaction, else
rolls back the transaction, if one is open.
- docs: add docstrings for important stuff
- deps: make dependency requirements more lenient